half-api approach with CustomLazyParamterWithCustomizableEvaluationLogic
