+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/expression/ExpressionEvaluator.scala
@@ -3,8 +3,7 @@ package pl.touk.nussknacker.engine.expression
 import pl.touk.nussknacker.engine.api.typed.CustomNodeValidationException
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
-import pl.touk.nussknacker.engine.compiledgraph.{BaseCompiledParameter, CompiledParameter}
-import pl.touk.nussknacker.engine.expression.parse.CompiledExpression
+import pl.touk.nussknacker.engine.compiledgraph.CompiledParameter
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.engine.variables.GlobalVariablesPreparer
 
@@ -39,7 +38,7 @@ class ExpressionEvaluator(
     globalVariablesPreparer: GlobalVariablesPreparer,
     listeners: Seq[ProcessListener],
     cacheGlobalVariables: Boolean
-) {
+) extends BaseExpressionEvaluator {
   private def prepareGlobals(jobData: JobData): Map[String, Any] =
     globalVariablesPreparer.prepareGlobalVariables(jobData).mapValuesNow(_.obj)
