+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/expression/parse/CompiledExpression.scala
@@ -1,18 +1,9 @@
 package pl.touk.nussknacker.engine.expression.parse
 
-import pl.touk.nussknacker.engine.api.Context
+import pl.touk.nussknacker.engine.api.{CompiledExpression, Context}
 import pl.touk.nussknacker.engine.api.expression.ExpressionTypingInfo
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.graph.expression.Expression.Language
-
-trait CompiledExpression {
-
-  def language: Language
-
-  def original: String
-
-  def evaluate[T](ctx: Context, globals: Map[String, Any]): T
-}
 
 sealed trait TypedValue
