+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/expression/parse/ExpressionParser.scala
@@ -1,6 +1,7 @@
 package pl.touk.nussknacker.engine.expression.parse
 
 import cats.data.ValidatedNel
+import pl.touk.nussknacker.engine.api.CompiledExpression
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
